Write scalars without type information (#1499)
NumPy 2.x changed the representation of scalaras including type
information. That lead to failing tests, since the test files were
written with the new representation (e.g. `np.int64(0)`).

This reverts to the old behavior of using plain `0`. Though, the actual
fix is likely to use `str()` or `f"{scalar!s}"` as described in the
release notes.